+++ dms/make_notification_requested_run_batch_events.sql
@@ -1,25 +1,26 @@
 --
-CREATE OR REPLACE PROCEDURE public.make_notification_requested_run_batch_events
-(
-    _infoOnly boolean = false,
-    _showDebug boolean = false
-)
-LANGUAGE plpgsql
-AS $$
+-- Name: make_notification_requested_run_batch_events(boolean, boolean, boolean, timestamp without time zone); Type: PROCEDURE; Schema: public; Owner: d3l243
+--
+
+CREATE OR REPLACE PROCEDURE public.make_notification_requested_run_batch_events(IN _infoonly boolean DEFAULT false, IN _showdebug boolean DEFAULT false, IN _deleteoldevents boolean DEFAULT true, IN _timestampoverride timestamp without time zone DEFAULT NULL::timestamp without time zone)
+    LANGUAGE plpgsql
+    AS $$
 /****************************************************
 **
 **  Desc:
-**      Add requested run batch notification events to notification event table
+**      Add requested run batch notification events to the notification event table
 **
 **  Arguments:
-**    _infoOnly     When true, show the number of notification events that would be added
-**    _showDebug    When _infoOnly is true, if _showDebug is true, show details on the events that would be added
+**    _infoOnly             When true, show the number of notification events that would be added
+**    _showDebug            When _infoOnly is true, if _showDebug is true, show details on the events that would be added
+**    _deleteOldEvents      When true (the default), delete requested run batch notification events that are more than 7 days old
+**    _timestampOverride    Optional, specific timestamp to use for finding events when _infoOnly is true; when _infoOnly is false, this procedure uses the current timestamp
 **
 **  Auth:   grk
 **  Date:   03/26/2010 grk - Initial version
 **          03/30/2010 grk - Added intermediate table
 **          04/01/2010 grk - Added Latest_Suspect_Dataset
-**          12/15/2024 mem - Ported to PostgreSQL
+**          02/14/2024 mem - Ported to PostgreSQL
 **
 *****************************************************/
 DECLARE
@@ -30,35 +31,40 @@ DECLARE
     _past timestamp;
     _future timestamp;
     _eventType int;
+    _formatSpecifier text;
     _batchInfo record;
     _eventInfo record;
 BEGIN
 
-    _infoOnly := Coalesce(_infoOnly, false);
-    _showDebug := Coalesce(_showDebug, false);
+    _infoOnly        := Coalesce(_infoOnly, false);
+    _showDebug       := Coalesce(_showDebug, false);
+    _deleteOldEvents := Coalesce(_deleteOldEvents, true);
+
+    If _infoOnly Then
+        _now := Coalesce(_timestampOverride, CURRENT_TIMESTAMP);
+    Else
+        If Not _timestampOverride Is Null Then
+            RAISE INFO 'Ignoring _timestampOverride since _infoOnly is false';
+        End If;
+
+        _now := CURRENT_TIMESTAMP;
+    End If;
 
     ---------------------------------------------------
     -- Window for requested run activity
     ---------------------------------------------------
 
-    _window := CURRENT_TIMESTAMP - INTERVAL '7 days'
+    _window := _now - INTERVAL '7 days';
 
     ---------------------------------------------------
     -- Window for batch creation date
     ---------------------------------------------------
 
-    _threshold := CURRENT_TIMESTAMP - INTERVAL '365 days'
-
-    ---------------------------------------------------
-    -- Earlier than batch creation window
-    -- (default for datasets with null start time)
-    ---------------------------------------------------
-
-    _now := CURRENT_TIMESTAMP;
+    _threshold := _now - INTERVAL '365 days';
 
     _past := make_date(2000, 1, 1);
 
-    _future := CURRENT_TIMESTAMP + INTERVAL '3 months';
+    _future := _now + INTERVAL '3 months';
 
     ---------------------------------------------------
     -- Temp table for batches of interest
@@ -82,51 +88,51 @@ BEGIN
     INSERT INTO Tmp_RequestedRunBatches
     SELECT RRB.Batch_ID,
            COUNT(RR.request_id) AS Num_Requests,
-           SUM(CASE
-                   WHEN TD.Dataset_ID IS NULL THEN 0
-                   ELSE 1
+           SUM(CASE WHEN TD.Dataset_ID IS NULL THEN 0
+                    ELSE 1
                END) AS Num_Datasets,
-           SUM(CASE
-                   WHEN TD.Acq_Time_Start IS NULL THEN 0
-                   ELSE 1
+           SUM(CASE WHEN TD.Acq_Time_Start IS NULL THEN 0
+                    ELSE 1
                END) AS Num_Datasets_With_Start_Time,
            MIN(Coalesce(TD.Created, _future)) AS Earliest_Dataset,
            MAX(Coalesce(TD.Created, _past)) AS Latest_Dataset,
-           MAX(CASE
-                   WHEN TD.dataset_rating_id BETWEEN - 5 AND - 1 THEN TD.Created
-                   ELSE _past
+           MAX(CASE WHEN TD.dataset_rating_id BETWEEN - 5 AND - 1 THEN TD.Created
+                    ELSE _past
                END) AS Latest_Suspect_Dataset
     FROM t_requested_run_batches AS RRB
          INNER JOIN t_requested_run AS RR
            ON RR.batch_id = RRB.Batch_ID
          LEFT OUTER JOIN t_dataset AS TD
-           ON TD.Dataset_ID = RR.DatasetID
+           ON TD.Dataset_ID = RR.Dataset_ID
     WHERE RRB.batch_id <> 0 AND
           RRB.created > _threshold
     GROUP BY RRB.batch_id;
 
     If _showDebug Then
+        RAISE INFO '';
+
+        _formatSpecifier := '%-10s %-12s %-12s %-20s %-20s %-20s %-22s';
 
         RAISE INFO '%',
-                    format('%-10s %-12s %-12s %-20s %-20s %-20s %-20s',
-                            'Batch ID',
-                            'Num Requests',
-                            'Num Datasets',
-                            'Num With Start Time',
-                            'Earliest Dataset',
-                            'Latest Dataset',
-                            'Latest Suspect Dataset'
+                    format(_formatSpecifier,
+                           'Batch ID',
+                           'Num Requests',
+                           'Num Datasets',
+                           'Num With Start Time',
+                           'Earliest Dataset',
+                           'Latest Dataset',
+                           'Latest Suspect Dataset'
                           );
 
         RAISE INFO '%',
-                    format('%-10s %-12s %-12s %-20s %-20s %-20s %-20s',
-                            '----------',
-                            '------------',
-                            '------------',
-                            '--------------------',
-                            '--------------------',
-                            '--------------------',
-                            '--------------------'
+                    format(_formatSpecifier,
+                           '----------',
+                           '------------',
+                           '------------',
+                           '--------------------',
+                           '--------------------',
+                           '--------------------',
+                           '----------------------'
                           );
 
         FOR _batchInfo IN
@@ -141,13 +147,13 @@ BEGIN
             ORDER BY Batch_ID
         LOOP
             RAISE INFO '%',
-                        format('%-10s %-12s %-12s %-20s %-20s %-20s %-20s',
+                        format(_formatSpecifier,
                                _batchInfo.BatchID,
                                _batchInfo.NumRequests,
                                _batchInfo.NumDatasets,
                                _batchInfo.NumWithStartTime,
-                               to_char(_batchInfo.EarliestDataset, 'yyyy-mm-dd hh12:mi AM'),
-                               to_char(_batchInfo.LatestDataset, 'yyyy-mm-dd hh12:mi AM'),
+                               to_char(_batchInfo.EarliestDataset,      'yyyy-mm-dd hh12:mi AM'),
+                               to_char(_batchInfo.LatestDataset,        'yyyy-mm-dd hh12:mi AM'),
                                to_char(_batchInfo.LatestSuspectDataset, 'yyyy-mm-dd hh12:mi AM')
                               );
 
@@ -160,72 +166,72 @@ BEGIN
 
     CREATE TEMP TABLE Tmp_NewEvents (
         Target_ID int,
-        Event_Type int
+        Event_Type_ID int
     );
 
     ---------------------------------------------------
     -- Event 'Requested Run Batch Start'
-    -- * Num_Datasets > 0
-    -- * Earliest_Dataset within window
-    -- * 'Requested Run Batch Start' event and RRB.ID not already in event table
+    --   Num_Datasets > 0
+    --   Earliest_Dataset within window
+    --   'Requested Run Batch Start' event and RRB.Batch_ID not already in event table
     ---------------------------------------------------
 
     _eventType := 1;
 
     INSERT INTO Tmp_NewEvents( target_id,
                                event_type_id )
-    SELECT entry_id,
+    SELECT Batch_ID,
            _eventType
-    FROM Tmp_AnalysisJobs
+    FROM Tmp_RequestedRunBatches
     WHERE Num_Datasets > 0 AND
           Earliest_Dataset BETWEEN _window AND _now AND
-          NOT EXISTS ( SELECT *
+          NOT EXISTS ( SELECT 1
                        FROM t_notification_event TNE
-                       WHERE TNE.target_id = Tmp_AnalysisJobs.entry_id
+                       WHERE TNE.target_id = Tmp_RequestedRunBatches.Batch_ID
                              AND
                              TNE.event_type_id = _eventType );
 
     ---------------------------------------------------
     -- Event 'Requested Run Batch Finish'
-    --- Num_Requests = Num_Datasets
-    --- Latest_Dataset within window
-    --- 'Requested Run Batch Finish' event and RRB.ID not already in event table
+    --   Num_Requests = Num_Datasets
+    --   Latest_Dataset within window
+    --   'Requested Run Batch Finish' event and RRB.Batch_ID not already in event table
     ---------------------------------------------------
 
     _eventType := 2;
 
     INSERT INTO Tmp_NewEvents( target_id,
                                event_type_id )
-    SELECT entry_id,
+    SELECT Batch_ID,
            _eventType
-    FROM Tmp_AnalysisJobs
+    FROM Tmp_RequestedRunBatches
     WHERE Num_Datasets = Num_Requests AND
           Latest_Dataset BETWEEN _window AND _now AND
-          NOT EXISTS ( SELECT *
+          NOT EXISTS ( SELECT 1
                        FROM t_notification_event TNE
-                       WHERE TNE.target_id = Tmp_AnalysisJobs.entry_id
+                       WHERE TNE.target_id = Tmp_RequestedRunBatches.Batch_ID
                              AND
                              TNE.event_type_id = _eventType );
 
     ---------------------------------------------------
     -- Event 'Requested Run Batch Acq Time Ready'
-    --- Num_Requests = Num_Datasets_With_Start_Time
-    --- Latest_Dataset within window
-    --- 'Requested Run Batch Acq Time Ready' event and RRB.ID not already in event table
+    --   Num_Requests = Num_Datasets_With_Start_Time
+    --   Latest_Dataset within window
+    --   'Requested Run Batch Acq Time Ready' event and RRB.Batch_ID not already in event table
     ---------------------------------------------------
 
     _eventType := 3;
 
     INSERT INTO Tmp_NewEvents( target_id,
                                event_type_id )
-    SELECT entry_id,
+    SELECT Batch_ID,
            _eventType
-    FROM Tmp_AnalysisJobs
+    FROM Tmp_RequestedRunBatches
     WHERE Num_Requests = Num_Datasets_With_Start_Time AND
           Latest_Dataset BETWEEN _window AND _now AND
-          NOT EXISTS ( SELECT *
+          NOT EXISTS ( SELECT 1
                        FROM t_notification_event TNE
-                       WHERE TNE.target_id = Tmp_AnalysisJobs.entry_id AND
+                       WHERE TNE.target_id = Tmp_RequestedRunBatches.Batch_ID AND
                              TNE.event_type_id = _eventType );
 
     If _infoOnly Then
@@ -234,18 +240,26 @@ BEGIN
         INTO _eventCount
         FROM Tmp_NewEvents;
 
+        If _showDebug Then
+            RAISE INFO '';
+        End If;
+
         RAISE INFO 'Would add % % to t_notification_event', _eventCount, public.check_plural(_eventCount, 'row', 'rows');
 
         If _showDebug Then
-            RAISE INFO '%',
-                        format('%-10s %-15s',
-                                'Batch ID',
-                                'Event Type ID');
+            RAISE INFO '';
+
+            _formatSpecifier := '%-10s %-15s';
 
             RAISE INFO '%',
-                        format('%-10s %-15s',
-                                '----------',
-                                '---------------');
+                        format(_formatSpecifier,
+                               'Batch ID',
+                               'Event Type ID');
+
+            RAISE INFO '%',
+                        format(_formatSpecifier,
+                               '----------',
+                               '---------------');
 
             FOR _eventInfo IN
                 SELECT target_id As TargetID,
@@ -254,9 +268,9 @@ BEGIN
                 ORDER BY target_id
             LOOP
                 RAISE INFO '%',
-                           format('%-10s %-15s',
-                                    _eventInfo.TargetID,
-                                    _eventInfo.EventTypeID);
+                           format(_formatSpecifier,
+                                  _eventInfo.TargetID,
+                                  _eventInfo.EventTypeID);
             END LOOP;
         End If;
 
@@ -275,18 +289,31 @@ BEGIN
                            WHERE TNE.target_id = Tmp_NewEvents.target_id AND
                                  TNE.event_type_id = Tmp_NewEvents.event_type_id );
 
-        ---------------------------------------------------
-        -- Clean out batch events older than window
-        ---------------------------------------------------
+        If _deleteOldEvents Then
 
-        DELETE FROM t_notification_event
-        WHERE event_type_id IN (1, 2, 3) AND
-              entered < _window;
+            ---------------------------------------------------
+            -- Clean out requested run batch events older than window
+            ---------------------------------------------------
+
+            DELETE FROM t_notification_event
+            WHERE event_type_id IN (1, 2, 3) AND
+                  entered < _window;
+
+        End If;
+
     End If;
 
-    DROP TABLE Tmp_AnalysisJobs;
+    DROP TABLE Tmp_RequestedRunBatches;
     DROP TABLE Tmp_NewEvents;
 END
 $$;
 
-COMMENT ON PROCEDURE public.make_notification_requested_run_batch_events IS 'MakeNotificationRequestedRunBatchEvents';
+
+ALTER PROCEDURE public.make_notification_requested_run_batch_events(IN _infoonly boolean, IN _showdebug boolean, IN _deleteoldevents boolean, IN _timestampoverride timestamp without time zone) OWNER TO d3l243;
+
+--
+-- Name: PROCEDURE make_notification_requested_run_batch_events(IN _infoonly boolean, IN _showdebug boolean, IN _deleteoldevents boolean, IN _timestampoverride timestamp without time zone); Type: COMMENT; Schema: public; Owner: d3l243
+--
+
+COMMENT ON PROCEDURE public.make_notification_requested_run_batch_events(IN _infoonly boolean, IN _showdebug boolean, IN _deleteoldevents boolean, IN _timestampoverride timestamp without time zone) IS 'MakeNotificationRequestedRunBatchEvents';
+
